WIP to add early concept of "collection" to codebase
