+++ content/collections/framework-fieldguide/index.md
@@ -1,0 +1,18 @@
+---
+{
+  title: "The Framework Field Guide",
+  associatedSeries: "Web Components 101",
+  description: "It's a booky book of book preportions",
+  authors: ["crutchcorn"],
+  buttons:
+    [
+      { text: "Read now", url: "/posts/webb-components-101-history" },
+      {
+        text: "Buy a physical copy",
+        url: "/posts/webb-components-101-history",
+      },
+    ],
+}
+---
+
+This is some text to talk about why the book is awesome - it just is
