+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from "./LicenseInfo";
+export * from "./CollectionInfo";
 export * from "./PostInfo";
 export * from "./PronounInfo";
 export * from "./RolesInfo";
